Move identity card
