+++ services/agora/src/pages/user-profile/opinions/index.vue
@@ -78,6 +78,7 @@ import ZKHoverEffect from "src/components/ui-library/ZKHoverEffect.vue";
 import CommentActionOptions from "src/components/post/comments/group/item/CommentActionOptions.vue";
 import CommentModeration from "src/components/post/comments/group/item/CommentModeration.vue";
 import HtmlContent from "src/components/post/display/HtmlContent.vue";
+import UserIdentity from "src/components/features/user/UserIdentityCard.vue";
 
 const { loadMoreUserComments, loadUserProfile } = useUserStore();
 const { profileData } = storeToRefs(useUserStore());
